chore(chromecast): Move cast state sync upwards
- Sync cast state regardless if the player is mounted or not
- Allow playback if cast is connected before player is rendered
